+++ n64/math/include/math.h
@@ -17,7 +17,10 @@
 #ifndef MATH_H_GUARD
 #define MATH_H_GUARD
 
+#include <ints.h>
+
 #include <math/mathf.h>
 #include <math/maths.h>
+#include <math/mtx.h>
 
 #endif